Implemented item stacking.
Implemented item stacking.

Also blacklisted an item from being dropped. This is apparently the item
that drops for faction tokens, and the server will change it to a
different item for your faction if you are in one.

Since we are still in VE, I have blacklisted it for the time being