Critical Security Patch + Minor Changes
A security issue was discovered by Raine Bannister. The nature of the security issue was SQL Injection, which made it possible for users to delete, edit, or create data in the central database, directly from the login screen. CDClient editing was also theoretically possible from the /play command, but I was unable to reproduce it.

Please update immediately to this version of the server. Any versions prior to this one are vulnerable to SQL Injection and are not secure!

Relevant issues:
https://github.com/cuzitsjonny/JALUS/issues/5
https://github.com/cuzitsjonny/JALUS/issues/6

Other changes:
- Blacklisted some inappropriate emotes.
- Added 2 new aliases for /play: /e and /emote.
- The wisp lee hand effect is now default off. You can toggle it using /wisp.
     - Reason for this is that some people are losing connection, and I suspect this is the cause. This also cleans up the client log a bit.
- Other changes.